+++ Plugin.cs
@@ -12,7 +12,7 @@ namespace StoreAndCraft
     {
         public const string ModGuid = "com.morda.storeandcraft";
         public const string ModName = "StoreAndCraft";
-        public const string ModVersion = "1.2.2";
+        public const string ModVersion = "1.2.3";
         public const string ModAuthor = "Morda";
 
         internal static Plugin Instance { get; private set; }